Fix for grid not rendering
Fix for grid not rendering on Firefox and IE (latest version) when using 'dateTime' and 'dateTimeTwoLines' renders. When converting the value in a Date object, FF was returning an "Invalid Date" due to the use of - instead of / as a valid date string across all browsers (Google Chrome is more permissive on this).