Add More Flexible Link Renderer (#307)
Make it possible to place a link on any field (including tvs), rather than restricting links to pagetitle only. I often use tvs for custom titling and the ability to group related custom data logically within tabs using manager customization rules, which prompted this proposed addition to the renderers. If you agree and the new method is acceptible, it could be used as the basis for altering the other methods that add links.